+++ Libary/src/main/java/com/bloodspy/spring/libary/controller/BookController.java
@@ -1,6 +1,9 @@
 package com.bloodspy.spring.libary.controller;
 
 import com.bloodspy.spring.libary.entity.BookEntity;
+import com.bloodspy.spring.libary.exceptionHandler.exceptions.NoSuchException;
+import com.bloodspy.spring.libary.returnMessage.ContainerReturnMessage;
+import com.bloodspy.spring.libary.returnMessage.ReturnMessageHandler;
 import com.bloodspy.spring.libary.service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -10,12 +13,18 @@ import java.util.List;
 @RestController
 @RequestMapping("/")
 public class BookController {
+    private final String entityName = "Book";
+
     @Autowired
-    public BookController(BookService bookService) {
+    public BookController(BookService bookService,
+                          ReturnMessageHandler returnMessageHandler) {
         this.bookService = bookService;
+        this.returnMessageHandler = returnMessageHandler;
     }
 
     BookService bookService;
+
+    ReturnMessageHandler returnMessageHandler;
 
     @GetMapping("/books")
     public List<BookEntity> getAllBook() {
@@ -28,27 +37,37 @@ public class BookController {
     public BookEntity getBook(@PathVariable(name = "id") int id) {
         BookEntity book = bookService.getBook(id);
 
+        if(book == null) {
+            throw new NoSuchException(entityName, id);
+        }
+
         return book;
     }
 
     @PostMapping("/books")
-    public String addBook(@RequestBody BookEntity book) {
+    public ContainerReturnMessage addBook(@RequestBody BookEntity book) {
        bookService.saveBook(book);
 
-       return "Book with id " + book.getId() + " was saved";
+       return returnMessageHandler.getAddMessage(entityName, book.getId());
     }
 
     @PutMapping("/books")
-    public String updateBook(@RequestBody BookEntity book) {
+    public ContainerReturnMessage updateBook(@RequestBody BookEntity book) {
         bookService.saveBook(book);
 
-        return "Book with id " + book.getId() + " was updated";
+        return returnMessageHandler.getUpdateMessage(entityName, book.getId());
     }
 
     @DeleteMapping("/books/{id}")
-    public String deleteBook(@PathVariable(name = "id") int id) {
+    public ContainerReturnMessage deleteBook(@PathVariable(name = "id") int id) {
+        BookEntity book = bookService.getBook(id);
+
+        if(book == null) {
+            throw new NoSuchException(entityName, id);
+        }
+
         bookService.deleteBook(id);
 
-        return "Book with id " + id + " was deleted";
+        return returnMessageHandler.getDeleteMessage(entityName, book.getId());
     }
 }
